+++ servers/defaults.go
@@ -89,13 +89,14 @@ var cmdAPI = map[string]cmdspecs{
 }
 
 const maxsensors = 200               // maximum number of allowed processors
-const mindelayrefusedconnection = 30 // mininum delay for refused connection
-var errormngt = [3]int{1, 5, 15}     // [min penalty, max panelty, max number of consecucitve errors]
-var tcpTokens chan bool              // token for accepting a TCP erquest
+const mindelayrefusedconnection = 30 // minimum delay for refused connection
+var errormngt = [3]int{1, 5, 15}     // [min penalty, max penalty, max number of consecutive errors]
+var tcpTokens chan bool              // token for accepting a TCP request
 var Kswitch bool                     // kill switch flag
 var RepCon bool                      // enables reporting on current
 
 // debug access control
-//var dbgMutex = &sync.Mutex{}     // lock to dbgRegistry
-//var dbgRegistry map[string]int64 // registry of curretly authorised IPs
-//const authInterval = 1           // autorisation interval for debug access in minutes
+var dbgMutex = &sync.RWMutex{}   // lock to dbgRegistry
+var dbgRegistry map[string]int64 // registry of currently authorised IPs
+const authDbgInterval = 10       // authorisation interval for debug access in minutes
+const pindbg = "pippopluto"      // debug pin